requirements.txt
tensorflow
torch
scikit-learn
xgboost
transformers
pandas
matplotlib
scipy
msclap
pydub
librosa
soundfile
pyyaml
tqdm_multiprocess
h5py
